refactor: add pipeline concept to OTLP traces and remove OTLP over gRPC
