+++ src/pipeline/src/etl.rs
@@ -20,12 +20,9 @@ pub mod processor;
 pub mod transform;
 pub mod value;
 
-use std::sync::Arc;
-
 use error::{
     IntermediateKeyIndexSnafu, PrepareValueMustBeObjectSnafu, YamlLoadSnafu, YamlParseSnafu,
 };
-use itertools::Itertools;
 use processor::{Processor, Processors};
 use snafu::{ensure, OptionExt, ResultExt};
 use transform::{Transformer, Transforms};
@@ -34,7 +31,6 @@ use yaml_rust::YamlLoader;
 
 use crate::dispatcher::{Dispatcher, Rule};
 use crate::etl::error::Result;
-use crate::{GreptimeTransformer, PipelineVersion};
 
 const DESCRIPTION: &str = "description";
 const PROCESSORS: &str = "processors";
@@ -212,57 +208,6 @@ pub(crate) fn find_key_index(intermediate_keys: &[String], key: &str, kind: &str
         .iter()
         .position(|k| k == key)
         .context(IntermediateKeyIndexSnafu { kind, key })
-}
-
-/// SelectInfo is used to store the selected keys from OpenTelemetry record attrs
-/// The key is used to uplift value from the attributes and serve as column name in the table
-#[derive(Default)]
-pub struct SelectInfo {
-    pub keys: Vec<String>,
-}
-
-/// Try to convert a string to SelectInfo
-/// The string should be a comma-separated list of keys
-/// example: "key1,key2,key3"
-/// The keys will be sorted and deduplicated
-impl From<String> for SelectInfo {
-    fn from(value: String) -> Self {
-        let mut keys: Vec<String> = value.split(',').map(|s| s.to_string()).sorted().collect();
-        keys.dedup();
-
-        SelectInfo { keys }
-    }
-}
-
-impl SelectInfo {
-    pub fn is_empty(&self) -> bool {
-        self.keys.is_empty()
-    }
-}
-
-pub const GREPTIME_INTERNAL_IDENTITY_PIPELINE_NAME: &str = "greptime_identity";
-
-/// Enum for holding information of a pipeline, which is either pipeline itself,
-/// or information that be used to retrieve a pipeline from `PipelineHandler`
-pub enum PipelineDefinition {
-    Resolved(Arc<Pipeline<GreptimeTransformer>>),
-    ByNameAndValue((String, PipelineVersion)),
-    GreptimeIdentityPipeline,
-}
-
-impl PipelineDefinition {
-    pub fn from_name(name: &str, version: PipelineVersion) -> Self {
-        if name == GREPTIME_INTERNAL_IDENTITY_PIPELINE_NAME {
-            Self::GreptimeIdentityPipeline
-        } else {
-            Self::ByNameAndValue((name.to_owned(), version))
-        }
-    }
-}
-
-pub enum PipelineWay {
-    OtlpLogDirect(Box<SelectInfo>),
-    Pipeline(PipelineDefinition),
 }
 
 #[cfg(test)]
