+++ src/servers/src/otlp/logs.rs
@@ -32,7 +32,8 @@ use snafu::{ensure, ResultExt};
 use super::trace::attributes::OtlpAnyValue;
 use super::utils::{bytes_to_hex_string, key_value_to_jsonb};
 use crate::error::{
-    IncompatibleSchemaSnafu, PipelineTransformSnafu, Result, UnsupportedJsonDataTypeForTagSnafu,
+    IncompatibleSchemaSnafu, NotSupportedSnafu, PipelineTransformSnafu, Result,
+    UnsupportedJsonDataTypeForTagSnafu,
 };
 use crate::pipeline::run_pipeline;
 use crate::query_handler::PipelineHandlerRef;
@@ -98,6 +99,10 @@ pub async fn to_grpc_insert_requests(
             let insert_requests = RowInsertRequests { inserts };
             Ok((insert_requests, len))
         }
+        _ => NotSupportedSnafu {
+            feat: "Unsupported pipeline for logs",
+        }
+        .fail(),
     }
 }
